Expressions and augmented expressions
Support simple operations like `variable = expression` and `variable +=
expression` which store their result in a separate property. Disabled by
default, enable via constructor or a property of the parser. Variable
scope can now be replaced as well. Use this to keep a context in the
interpreter as demonstration.

Update README and unit tests.